End of semester QT-hist
I was able to add commas to the axis by editing the labelCuts function. I don't know if you want that to be changed for all graphs or just mine- I can re-name the function something else if you want to keep them separate. 

This function operates with the following constraints: can't have a quantile less than 1 or over 20. The automated options return between 5 and 100 bins and that value is found based on the size of the dataset. The automated quantile for the smallest data (n less than 500) is 5% on each side of the graph, and for the largest data (n over 10,000) the automated value is 1%.